src/voice/requirements.txt
azure-ai-voicelive[aiohttp]
azure-identity
python-dotenv
websockets
